requirements.txt
flask==3.0.0
flask-login==0.6.3
flask-socketio==5.3.6
python-dotenv==1.0.0
pymongo==4.6.1
bcrypt==4.1.2
simple-websocket==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
gunicorn==21.2.0
eventlet==0.35.2 